+++ src/App.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import './App.css';
+import './style/App.css';
 
 function App() {
   return (
